+++ modules/cloudfront/basic_auth_function/outputs.tf
@@ -1,0 +1,3 @@
+output "cf_function_arn" {
+  value = aws_cloudfront_function.basicauth_cf_function.arn
+}