+++ modules/cloudfront/custom_origin/inputs.tf
@@ -127,7 +127,7 @@ variable "auth_header_name" {
 variable "custom_http_whitelisted_headers" {
   description = "List of whitelisted http headers to have CloudFront forward to origin"
   type        = list(string)
-  default     = ["Authorization", "Origin", "Access-Control-Request-Headers", "Access-Control-Request-Method", "User-Agent"]
+  default     = ["Authorization", "Origin", "Access-Control-Request-Headers", "Access-Control-Request-Method", "User-Agent", "filmdrop-authorized"]
 }
 
 variable "custom_error_response" {
@@ -158,6 +158,54 @@ variable "whitelist_ips" {
   description = "List of ips to filter access for."
   type        = set(string)
   default     = []
+}
+
+variable "cf_function_name" {
+  description = "Name of the CF function"
+  type        = string
+  default     = ""
+}
+
+variable "cf_function_runtime" {
+  description = "CF function runtime"
+  type        = string
+  default     = "cloudfront-js-2.0"
+}
+
+variable "cf_function_code_path" {
+  description = "CF function code path"
+  type        = string
+  default     = ""
+}
+
+variable "attach_cf_function" {
+  description = "Should attach a function to CF or not"
+  type        = bool
+  default     = false
+}
+
+variable "cf_function_event_type" {
+  description = "Eventtype for the function"
+  type        = string
+  default     = "viewer-request"
+}
+
+variable "create_cf_function" {
+  description = "Should create a CF function or not"
+  type        = bool
+  default     = false
+}
+
+variable "create_cf_basicauth_function" {
+  description = "Should create the BasicAuth CF function or not"
+  type        = bool
+  default     = false
+}
+
+variable "cf_function_arn" {
+  description = "CF Function arn in case to get in input"
+  type        = string
+  default     = ""
 }
 
 variable "project_name" {