+++ modules/cloudfront/s3_website/s3_website.tf
@@ -27,12 +27,13 @@ module "cloudfront_distribution" {
   max_ttl                      = var.max_ttl
   domain_name                  = var.domain_name
   custom_error_response        = var.custom_error_response
-  attach_cf_function           = var.attach_cf_function
   cf_function_name             = var.cf_function_name
   cf_function_runtime          = var.cf_function_runtime
   cf_function_code_path        = var.cf_function_code_path
+  attach_cf_function           = var.attach_cf_function
   cf_function_event_type       = var.cf_function_event_type
   create_cf_function           = var.create_cf_function
+  create_cf_basicauth_function = var.create_cf_basicauth_function
   cf_function_arn              = var.cf_function_arn
   project_name                 = var.project_name
   environment                  = var.environment