+++ profiles/core/inputs.tf
@@ -107,6 +107,16 @@ variable "stac_server_inputs" {
     opensearch_ebs_volume_size                  = number
     stac_server_and_titiler_s3_arns             = list(string)
     web_acl_id                                  = string
+    auth_function = object({
+      cf_function_name             = string
+      cf_function_runtime          = string
+      cf_function_code_path        = string
+      attach_cf_function           = bool
+      cf_function_event_type       = string
+      create_cf_function           = bool
+      create_cf_basicauth_function = bool
+      cf_function_arn              = string
+    })
     ingest = object({
       source_catalog_url               = string
       destination_collections_list     = string
@@ -138,6 +148,16 @@ variable "stac_server_inputs" {
     opensearch_ebs_volume_size                  = 35
     stac_server_and_titiler_s3_arns             = []
     web_acl_id                                  = ""
+    auth_function = {
+      cf_function_name             = ""
+      cf_function_runtime          = "cloudfront-js-2.0"
+      cf_function_code_path        = ""
+      attach_cf_function           = false
+      cf_function_event_type       = "viewer-request"
+      create_cf_function           = false
+      create_cf_basicauth_function = false
+      cf_function_arn              = ""
+    }
     ingest = {
       source_catalog_url               = ""
       destination_collections_list     = ""
@@ -164,6 +184,16 @@ variable "titiler_inputs" {
     mosaic_titiler_waf_allowed_url  = string
     mosaic_titiler_host_header      = string
     web_acl_id                      = string
+    auth_function = object({
+      cf_function_name             = string
+      cf_function_runtime          = string
+      cf_function_code_path        = string
+      attach_cf_function           = bool
+      cf_function_event_type       = string
+      create_cf_function           = bool
+      create_cf_basicauth_function = bool
+      cf_function_arn              = string
+    })
   })
   default = {
     app_name                        = "titiler"
@@ -173,6 +203,16 @@ variable "titiler_inputs" {
     mosaic_titiler_waf_allowed_url  = ""
     mosaic_titiler_host_header      = ""
     web_acl_id                      = ""
+    auth_function = {
+      cf_function_name             = ""
+      cf_function_runtime          = "cloudfront-js-2.0"
+      cf_function_code_path        = ""
+      attach_cf_function           = false
+      cf_function_event_type       = "viewer-request"
+      create_cf_function           = false
+      create_cf_basicauth_function = false
+      cf_function_arn              = ""
+    }
   }
 }
 
@@ -184,6 +224,16 @@ variable "analytics_inputs" {
     jupyterhub_elb_acm_cert_arn = string
     jupyterhub_elb_domain_alias = string
     create_credentials          = bool
+    auth_function = object({
+      cf_function_name             = string
+      cf_function_runtime          = string
+      cf_function_code_path        = string
+      attach_cf_function           = bool
+      cf_function_event_type       = string
+      create_cf_function           = bool
+      create_cf_basicauth_function = bool
+      cf_function_arn              = string
+    })
   })
   default = {
     app_name                    = "analytics"
@@ -191,6 +241,16 @@ variable "analytics_inputs" {
     jupyterhub_elb_acm_cert_arn = ""
     jupyterhub_elb_domain_alias = ""
     create_credentials          = true
+    auth_function = {
+      cf_function_name             = ""
+      cf_function_runtime          = "cloudfront-js-2.0"
+      cf_function_code_path        = ""
+      attach_cf_function           = false
+      cf_function_event_type       = "viewer-request"
+      create_cf_function           = false
+      create_cf_basicauth_function = false
+      cf_function_arn              = ""
+    }
   }
 }
 
@@ -209,6 +269,16 @@ variable "console_ui_inputs" {
     filmdrop_ui_config_file = string
     filmdrop_ui_logo_file   = string
     filmdrop_ui_logo        = string
+    auth_function = object({
+      cf_function_name             = string
+      cf_function_runtime          = string
+      cf_function_code_path        = string
+      attach_cf_function           = bool
+      cf_function_event_type       = string
+      create_cf_function           = bool
+      create_cf_basicauth_function = bool
+      cf_function_arn              = string
+    })
   })
   default = {
     app_name     = "console"
@@ -225,6 +295,16 @@ variable "console_ui_inputs" {
     filmdrop_ui_config_file = ""
     filmdrop_ui_logo_file   = ""
     filmdrop_ui_logo        = "bm9uZQo=" # Base64: 'none'
+    auth_function = {
+      cf_function_name             = ""
+      cf_function_runtime          = "cloudfront-js-2.0"
+      cf_function_code_path        = ""
+      attach_cf_function           = false
+      cf_function_event_type       = "viewer-request"
+      create_cf_function           = false
+      create_cf_basicauth_function = false
+      cf_function_arn              = ""
+    }
   }
 }
 
@@ -241,6 +321,16 @@ variable "cirrus_dashboard_inputs" {
     }))
     cirrus_api_endpoint_base = string
     cirrus_dashboard_release = string
+    auth_function = object({
+      cf_function_name             = string
+      cf_function_runtime          = string
+      cf_function_code_path        = string
+      attach_cf_function           = bool
+      cf_function_event_type       = string
+      create_cf_function           = bool
+      create_cf_basicauth_function = bool
+      cf_function_arn              = string
+    })
   })
   default = {
     app_name     = "dashboard"
@@ -255,6 +345,16 @@ variable "cirrus_dashboard_inputs" {
     ]
     cirrus_api_endpoint_base = ""
     cirrus_dashboard_release = "v0.5.1"
+    auth_function = {
+      cf_function_name             = ""
+      cf_function_runtime          = "cloudfront-js-2.0"
+      cf_function_code_path        = ""
+      attach_cf_function           = false
+      cf_function_event_type       = "viewer-request"
+      create_cf_function           = false
+      create_cf_basicauth_function = false
+      cf_function_arn              = ""
+    }
   }
 }